extracted share event function for reusability
